updated the message frontend
