+++ .env
@@ -4,12 +4,12 @@ DFX_VERSION='0.15.1'
 DFX_NETWORK='local'
 CANISTER_CANDID_PATH_message_board='/home/same/Desktop/message_board_frontend/backend/index.did'
 CANISTER_CANDID_PATH_MESSAGE_BOARD='/home/same/Desktop/message_board_frontend/backend/index.did'
-MESSAGE_BOARD_CANISTER_ID='b77ix-eeaaa-aaaaa-qaada-cai'
-CANISTER_ID_MESSAGE_BOARD='b77ix-eeaaa-aaaaa-qaada-cai'
-CANISTER_ID_message_board='b77ix-eeaaa-aaaaa-qaada-cai'
-FRONTEND_CANISTER_ID='bw4dl-smaaa-aaaaa-qaacq-cai'
-CANISTER_ID_FRONTEND='bw4dl-smaaa-aaaaa-qaacq-cai'
-CANISTER_ID_frontend='bw4dl-smaaa-aaaaa-qaacq-cai'
-CANISTER_ID='bw4dl-smaaa-aaaaa-qaacq-cai'
+MESSAGE_BOARD_CANISTER_ID='be2us-64aaa-aaaaa-qaabq-cai'
+CANISTER_ID_MESSAGE_BOARD='be2us-64aaa-aaaaa-qaabq-cai'
+CANISTER_ID_message_board='be2us-64aaa-aaaaa-qaabq-cai'
+FRONTEND_CANISTER_ID='bkyz2-fmaaa-aaaaa-qaaaq-cai'
+CANISTER_ID_FRONTEND='bkyz2-fmaaa-aaaaa-qaaaq-cai'
+CANISTER_ID_frontend='bkyz2-fmaaa-aaaaa-qaaaq-cai'
+CANISTER_ID='bkyz2-fmaaa-aaaaa-qaaaq-cai'
 CANISTER_CANDID_PATH='/home/same/Desktop/message_board_frontend/.dfx/local/canisters/frontend/assetstorage.did'
 # END DFX CANISTER ENVIRONMENT VARIABLES
