+++ package.json
@@ -17,6 +17,7 @@
     "postcss-loader": "^7.3.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.20.1",
     "uuid": "^9.0.1"
   },
   "devDependencies": {
